requirements.txt
fastapi
python-dotenv
uvicorn
pydantic
websockets
openai
itsdangerous
duckduckgo_search
boto3
Pillow
bs4